[refactor] 불필요한 클래스 삭제
- 게시글 수정 Application 삭제 -> PostService에 통합함
- 동일한 역할을 수행하는 Service가 2개나 존재하여 1개를 삭제하기로 판단함.
